Move issue #56 composer-focus tests to a dedicated file
The 189-line test addition pushed src/state/issues_tests_detail_flow.rs
to 1143 lines, over the CI hard limit of 1000. Move the five issue #56
regression tests into a new src/state/issues_tests_composer_focus.rs
module (with self-contained helpers) and register it in state/mod.rs.
No production behavior change; all tests still pass.
